+++ relayer-cli/src/commands/light.rs
@@ -2,12 +2,17 @@
 
 use abscissa_core::{Command, Options, Runnable};
 
-mod init;
+mod add;
+mod rm;
 
 /// `light` subcommand
 #[derive(Command, Debug, Options, Runnable)]
 pub enum LightCmd {
-    /// The `light init` subcommand
-    #[options(help = "initiate a light client for a given chain")]
-    Init(init::InitCmd),
+    /// The `light add` subcommand
+    #[options(help = "add a light client peer for a given chain")]
+    Add(add::AddCmd),
+
+    /// The `light rm` subcommand
+    #[options(help = "remove a light client peer for a given chain")]
+    Rm(rm::RmCmd),
 }